@@ -14,9 +14,9 @@
     "http-proxy-middleware": "^1.0.6",
     "morgan": "^1.10.0",
     "oidc-client": "^1.9.0",
-    "react": "^17.0.2",
+    "react": "^18.1.0",
     "react-bootstrap": "^1.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.1.0",
     "react-draft-wysiwyg": "^1.14.7",
     "react-helmet": "^6.1.0",
     "react-router": "^5.2.0",
@@ -36,9 +36,9 @@
     "@types/enzyme-adapter-react-16": "1.0.6",
     "@types/html-to-draftjs": "^1.4.0",
     "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.39",
+    "@types/react": "^18.0.9",
     "@types/react-bootstrap": "^1.0.1",
-    "@types/react-dom": "^17.0.11",
+    "@types/react-dom": "^18.0.4",
     "@types/react-draft-wysiwyg": "^1.13.2",
     "@types/react-helmet": "^6.1.5",
     "@types/react-router-dom": "^5.3.3",
