@@ -1,6 +1,6 @@
 import 'bootstrap/dist/css/bootstrap.css';
 import * as React from 'react';
-import * as ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 import './index.scss';
@@ -8,10 +8,12 @@ import './index.scss';
 const baseUrl = document.getElementsByTagName('base')[0].getAttribute('href');
 
 if (baseUrl) {
-  ReactDOM.render(
+  const container = document.getElementById('root');
+  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+  const root = createRoot(container!); // createRoot(container!) if you use TypeScript
+  root.render(
     <BrowserRouter basename={baseUrl}>
       <App />
-    </BrowserRouter>,
-    document.getElementById('root')
+    </BrowserRouter>
   );
 }
