@@ -107,9 +107,12 @@ function PostManager(props: PostManagerProps): JSX.Element {
     if (settings && markedPost) {
       setConfirmModal(false);
       const token = await authService.getAccessToken();
-      fetch(`clientapi/blog/deletepost?postid=${+markedPost.id}`, {
-        headers: !token ? {} : { Authorization: `Bearer ${token}` },
-      })
+      fetch(
+        `${settings?.baseHref}api/blog/deletepost?postid=${+markedPost.id}`,
+        {
+          headers: !token ? {} : { Authorization: `Bearer ${token}` },
+        }
+      )
         .then((response: { json: () => Promise<boolean> }) => response.json())
         .then(async (resp: boolean) => {
           if (resp === true) {
